test(Effect): consolidate scoped service restoration coverage
